feat: comprehensive static analysis improvements and testing enhancements
Major improvements to code quality, type safety, and testing coverage:

## Static Analysis & Code Quality
- Configure Psalm for strict type checking with baseline
- Update Rector configuration for PHP 8.3 best practices
- Add comprehensive PHPStan configuration
- Apply PHP CS Fixer code style improvements
- Add missing @throws annotations throughout codebase

## New Features & Interfaces
- Add OpenFgaGateInterface for better contract definition
- Implement PermissionCheckRequest DTO for type safety
- Enhance authorization service provider with strict typing
- Improve cache warmer with better error handling

## Testing Infrastructure
- Add comprehensive test suite for OpenFgaGate functionality
- Implement edge case testing for authorization scenarios
- Add middleware testing coverage
- Create boundary condition tests for robust error handling
- Add authorization error handling test suite
- Implement tagged cache testing

## Documentation
- Add comprehensive CHANGELOG.md
- Update README.md with CI status and testing information
- Document all major features and improvements

## CI/QA Status
✅ Tests: 263 passing, 1 skipped (744 assertions)
✅ PHPStan: No errors found
✅ Rector: No changes needed
✅ PHP CS Fixer: Code style compliant
⚠️ Psalm: 279 type-related errors (mostly mixed assignments)
📊 Coverage: 30.1% (working toward 90%+ target)

This represents a major step toward production readiness with
improved type safety, comprehensive testing, and proper CI integration.

+++ src/Console/Commands/CheckCommand.php
@@ -80,7 +80,7 @@ final class CheckCommand extends Command
 
             $allowed = $manager->check($user, $relation, $object, $contextualTuples, $context, $connection);
 
-            $duration = round((float) ((microtime(true) - $startTime) * 1000), 2);
+            $duration = round((microtime(true) - $startTime) * 1000.0, 2);
 
             if (true === $this->option('json')) {
                 $jsonOutput = json_encode([
